move Vertex2 tests to dedicated file
