+++ honeycomb-core/src/spatial_repr/tests.rs
@@ -64,11 +64,11 @@ mod coords {
 // --- vector
 
 mod vector {
-    // utils
     use super::almost_equal;
     use crate::{CoordsError, Vector2};
+    // utils
     macro_rules! almost_equal_vec {
-        ($lhs: expr, $rhs: expr, $t: ty) => {
+        ($lhs: expr, $rhs: expr) => {
             almost_equal($lhs.x(), $rhs.x()) & almost_equal($lhs.y(), $rhs.y())
         };
     }
@@ -93,23 +93,19 @@ mod vector {
                 let along_y = Vector2::<$t>::unit_y() * 3.0;
                 assert!(almost_equal_vec!(
                     along_x.unit_dir().unwrap(),
-                    Vector2::<$t>::unit_x(),
-                    $t
+                    Vector2::<$t>::unit_x()
                 ));
                 assert!(almost_equal_vec!(
                     Vector2::<$t>::unit_x().unit_dir().unwrap(),
-                    Vector2::<$t>::unit_x(),
-                    $t
+                    Vector2::<$t>::unit_x()
                 ));
                 assert!(almost_equal_vec!(
                     along_y.unit_dir().unwrap(),
-                    Vector2::<$t>::unit_y(),
-                    $t
+                    Vector2::<$t>::unit_y()
                 ));
                 assert!(almost_equal_vec!(
                     (along_x + along_y).unit_dir().unwrap(),
-                    Vector2::<$t>::from((4.0 / 5.0, 3.0 / 5.0)),
-                    $t
+                    Vector2::<$t>::from((4.0 / 5.0, 3.0 / 5.0))
                 ));
                 let origin: Vector2<$t> = Vector2::default();
                 assert_eq!(origin.unit_dir(), Err(CoordsError::InvalidUnitDir));
@@ -124,23 +120,19 @@ mod vector {
                 let along_y = Vector2::<$t>::unit_y() * 3.0;
                 assert!(almost_equal_vec!(
                     along_x.normal_dir().unwrap(),
-                    Vector2::<$t>::unit_y(),
-                    $t
+                    Vector2::<$t>::unit_y()
                 ));
                 assert!(almost_equal_vec!(
                     Vector2::<$t>::unit_x().normal_dir().unwrap(),
-                    Vector2::<$t>::unit_y(),
-                    $t
+                    Vector2::<$t>::unit_y()
                 ));
                 assert!(almost_equal_vec!(
                     along_y.normal_dir().unwrap(),
-                    -Vector2::<$t>::unit_x(),
-                    $t
+                    -Vector2::<$t>::unit_x()
                 ));
                 assert!(almost_equal_vec!(
                     Vector2::<$t>::unit_y().normal_dir().unwrap(),
-                    -Vector2::<$t>::unit_x(),
-                    $t
+                    -Vector2::<$t>::unit_x()
                 ));
                 let origin: Vector2<$t> = Vector2::default();
                 assert_eq!(origin.normal_dir(), Err(CoordsError::InvalidUnitDir));
@@ -160,4 +152,70 @@ mod vector {
 
 // --- vertex
 
-mod vertex {}
+mod vertex {
+    use crate::{Vector2, Vertex2};
+    // tests
+    #[test]
+    fn add_vertex_vector() {
+        {
+            let mut a: Vertex2<f64> = Vertex2::from((1.0, 1.0));
+            let b: Vector2<f64> = Vector2::from((1.0, 0.0));
+            let a_moved = a + b;
+            assert_eq!(a_moved, Vertex2::from((2.0, 1.0)));
+            a += &b;
+            assert_eq!(a, a_moved);
+            a += b;
+            assert_eq!(a, Vertex2::from((3.0, 1.0)));
+        }
+        {
+            let mut a: Vertex2<f32> = Vertex2::from((1.0, 1.0));
+            let b: Vector2<f32> = Vector2::from((1.0, 0.0));
+            let a_moved = a + b;
+            assert_eq!(a_moved, Vertex2::from((2.0, 1.0)));
+            a += &b;
+            assert_eq!(a, a_moved);
+            a += b;
+            assert_eq!(a, Vertex2::from((3.0, 1.0)));
+        }
+    }
+
+    #[test]
+    fn sub_vertex_vector() {
+        {
+            let mut a: Vertex2<f64> = Vertex2::from((1.0, 1.0));
+            let b: Vector2<f64> = Vector2::from((1.0, 0.0));
+            let a_moved = a - b;
+            assert_eq!(a_moved, Vertex2::from((0.0, 1.0)));
+            a -= &b;
+            assert_eq!(a, a_moved);
+            a -= b;
+            assert_eq!(a, Vertex2::from((-1.0, 1.0)));
+        }
+        {
+            let mut a: Vertex2<f32> = Vertex2::from((1.0, 1.0));
+            let b: Vector2<f32> = Vector2::from((1.0, 0.0));
+            let a_moved = a - b;
+            assert_eq!(a_moved, Vertex2::from((0.0, 1.0)));
+            a -= &b;
+            assert_eq!(a, a_moved);
+            a -= b;
+            assert_eq!(a, Vertex2::from((-1.0, 1.0)));
+        }
+    }
+
+    #[test]
+    fn sub_vertex_vertex() {
+        {
+            let a: Vertex2<f64> = Vertex2::from((1.0, 1.0));
+            let b: Vertex2<f64> = Vertex2::from((1.0, 0.0));
+            let ab = b - a;
+            assert_eq!(ab, Vector2::from((0.0, -1.0)));
+        }
+        {
+            let a: Vertex2<f32> = Vertex2::from((1.0, 1.0));
+            let b: Vertex2<f32> = Vertex2::from((1.0, 0.0));
+            let ab = b - a;
+            assert_eq!(ab, Vector2::from((0.0, -1.0)));
+        }
+    }
+}
