+++ honeycomb-core/src/spatial_repr/vertex.rs
@@ -249,74 +249,3 @@ impl<T: CoordsFloat> AttributeBind for Vertex2<T> {
         OrbitPolicy::Vertex
     }
 }
-
-// ------ TESTS
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn add_vertex_vector() {
-        {
-            let mut a: Vertex2<f64> = Vertex2::from((1.0, 1.0));
-            let b: Vector2<f64> = Vector2::from((1.0, 0.0));
-            let a_moved = a + b;
-            assert_eq!(a_moved, Vertex2::from((2.0, 1.0)));
-            a += &b;
-            assert_eq!(a, a_moved);
-            a += b;
-            assert_eq!(a, Vertex2::from((3.0, 1.0)));
-        }
-        {
-            let mut a: Vertex2<f32> = Vertex2::from((1.0, 1.0));
-            let b: Vector2<f32> = Vector2::from((1.0, 0.0));
-            let a_moved = a + b;
-            assert_eq!(a_moved, Vertex2::from((2.0, 1.0)));
-            a += &b;
-            assert_eq!(a, a_moved);
-            a += b;
-            assert_eq!(a, Vertex2::from((3.0, 1.0)));
-        }
-    }
-
-    #[test]
-    fn sub_vertex_vector() {
-        {
-            let mut a: Vertex2<f64> = Vertex2::from((1.0, 1.0));
-            let b: Vector2<f64> = Vector2::from((1.0, 0.0));
-            let a_moved = a - b;
-            assert_eq!(a_moved, Vertex2::from((0.0, 1.0)));
-            a -= &b;
-            assert_eq!(a, a_moved);
-            a -= b;
-            assert_eq!(a, Vertex2::from((-1.0, 1.0)));
-        }
-        {
-            let mut a: Vertex2<f32> = Vertex2::from((1.0, 1.0));
-            let b: Vector2<f32> = Vector2::from((1.0, 0.0));
-            let a_moved = a - b;
-            assert_eq!(a_moved, Vertex2::from((0.0, 1.0)));
-            a -= &b;
-            assert_eq!(a, a_moved);
-            a -= b;
-            assert_eq!(a, Vertex2::from((-1.0, 1.0)));
-        }
-    }
-
-    #[test]
-    fn sub_vertex_vertex() {
-        {
-            let a: Vertex2<f64> = Vertex2::from((1.0, 1.0));
-            let b: Vertex2<f64> = Vertex2::from((1.0, 0.0));
-            let ab = b - a;
-            assert_eq!(ab, Vector2::from((0.0, -1.0)));
-        }
-        {
-            let a: Vertex2<f32> = Vertex2::from((1.0, 1.0));
-            let b: Vertex2<f32> = Vertex2::from((1.0, 0.0));
-            let ab = b - a;
-            assert_eq!(ab, Vector2::from((0.0, -1.0)));
-        }
-    }
-}
